feat: refactor FallbackNotificationService to use NotifySettingsTable and create NotifySettingsTable class
