+++ surf-chat-fallback/src/main/kotlin/dev/slne/surf/chat/fallback/service/FallbackNotificationService.kt
@@ -2,11 +2,11 @@ package dev.slne.surf.chat.fallback.service
 
 import com.google.auto.service.AutoService
 import dev.slne.surf.chat.core.service.NotificationService
+import dev.slne.surf.chat.fallback.table.NotifySettingsTable
 import kotlinx.coroutines.Dispatchers
 import net.kyori.adventure.util.Services
 import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
-import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
 import org.jetbrains.exposed.sql.transactions.transaction
@@ -15,31 +15,22 @@ import java.util.*
 
 @AutoService(NotificationService::class)
 class FallbackNotificationService : NotificationService, Services.Fallback {
-    object NotificationSettings : Table("chat_notifications") {
-        val userUuid =
-            varchar("user_uuid", 36).transform({ UUID.fromString(it) }, { it.toString() })
-        val pingsEnabled = bool("pings_enabled").default(true)
-        val invitesEnabled = bool("invites_enabled").default(true)
-
-        override val primaryKey = PrimaryKey(userUuid)
-    }
-
     override fun createTable() {
         transaction {
-            SchemaUtils.create(NotificationSettings)
+            SchemaUtils.create(NotifySettingsTable)
         }
     }
 
     override suspend fun pingsEnabled(uuid: UUID) =
         newSuspendedTransaction(Dispatchers.IO) {
-            NotificationSettings.selectAll().where(NotificationSettings.userUuid eq uuid)
+            NotifySettingsTable.selectAll().where(NotifySettingsTable.userUuid eq uuid)
                 .firstOrNull()?.let {
-                    it[NotificationSettings.pingsEnabled]
+                    it[NotifySettingsTable.pingsEnabled]
                 } ?: true
         }
 
     override suspend fun enablePings(uuid: UUID) = newSuspendedTransaction(Dispatchers.IO) {
-        NotificationSettings.upsert {
+        NotifySettingsTable.upsert {
             it[userUuid] = uuid
             it[pingsEnabled] = true
         }
@@ -47,7 +38,7 @@ class FallbackNotificationService : NotificationService, Services.Fallback {
     }
 
     override suspend fun disablePings(uuid: UUID) = newSuspendedTransaction(Dispatchers.IO) {
-        NotificationSettings.upsert {
+        NotifySettingsTable.upsert {
             it[userUuid] = uuid
             it[pingsEnabled] = false
         }
@@ -56,14 +47,14 @@ class FallbackNotificationService : NotificationService, Services.Fallback {
     }
 
     override suspend fun invitesEnabled(uuid: UUID) = newSuspendedTransaction(Dispatchers.IO) {
-        NotificationSettings.selectAll().where(NotificationSettings.userUuid eq uuid)
+        NotifySettingsTable.selectAll().where(NotifySettingsTable.userUuid eq uuid)
             .firstOrNull()?.let {
-                it[NotificationSettings.invitesEnabled]
+                it[NotifySettingsTable.invitesEnabled]
             } ?: true
     }
 
     override suspend fun enableInvites(uuid: UUID) = newSuspendedTransaction(Dispatchers.IO) {
-        NotificationSettings.upsert {
+        NotifySettingsTable.upsert {
             it[userUuid] = uuid
             it[invitesEnabled] = true
         }
@@ -72,7 +63,7 @@ class FallbackNotificationService : NotificationService, Services.Fallback {
 
 
     override suspend fun disableInvites(uuid: UUID) = newSuspendedTransaction(Dispatchers.IO) {
-        NotificationSettings.upsert {
+        NotifySettingsTable.upsert {
             it[userUuid] = uuid
             it[invitesEnabled] = false
         }
